refactor: flatten random module
